# docs/openapi.yaml
openapi: 3.0.3
info:
  title: KubeNova API
  version: 0.2.0
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Tenant:
      type: object
      properties:
        name: { type: string }
        owners: { type: array, items: { type: string } }
        labels: { type: object, additionalProperties: { type: string } }
        annotations: { type: object, additionalProperties: { type: string } }
        createdAt: { type: string, format: date-time }
      required: [name]
    Project:
      type: object
      properties:
        tenant: { type: string }
        name: { type: string }
        labels: { type: object, additionalProperties: { type: string } }
        annotations: { type: object, additionalProperties: { type: string } }
        quota: { type: object, additionalProperties: { type: string } }
        createdAt: { type: string, format: date-time }
      required: [tenant, name]
    App:
      type: object
      properties:
        tenant: { type: string }
        project: { type: string }
        name: { type: string }
        image: { type: string }
        properties: { type: object }
        createdAt: { type: string, format: date-time }
      required: [tenant, project, name]
    KubeconfigGrant:
      type: object
      properties:
        tenant: { type: string }
        project: { type: string }
        role: { type: string, enum: [tenant-admin, tenant-dev, read-only] }
        expires: { type: string, format: date-time }
        kubeconfig: { type: string, format: byte }
      required: [tenant, role]
paths:
  /healthz:
    get:
      responses: { '200': { description: ok } }
  /metrics:
    get:
      responses: { '200': { description: prometheus metrics } }
  /sync/events:
    post:
      responses: { '204': { description: accepted } }
  /sync/metrics:
    post:
      responses: { '204': { description: accepted } }
  /sync/logs:
    post:
      responses: { '204': { description: accepted } }
  /api/v1/tenants:
    get:
      security: [ { bearerAuth: [] } ]
      responses:
        '200':
          description: list tenants
          content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Tenant' } } } }
    post:
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/Tenant' } } }
      responses: { '200': { description: created } }
  /api/v1/tenants/{name}:
    get:
      security: [ { bearerAuth: [] } ]
      parameters: [ { name: name, in: path, required: true, schema: { type: string } } ]
      responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/Tenant' } } } } }
    put:
      security: [ { bearerAuth: [] } ]
      parameters: [ { name: name, in: path, required: true, schema: { type: string } } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/Tenant' } } } }
      responses: { '200': { description: ok } }
    delete:
      security: [ { bearerAuth: [] } ]
      parameters: [ { name: name, in: path, required: true, schema: { type: string } } ]
      responses: { '200': { description: ok } }
  /api/v1/projects:
    post:
      security: [ { bearerAuth: [] } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
      responses: { '200': { description: ok } }
  /api/v1/projects/{tenant}/{name}:
    get:
      security: [ { bearerAuth: [] } ]
      parameters:
        - { name: tenant, in: path, required: true, schema: { type: string } }
        - { name: name, in: path, required: true, schema: { type: string } }
      responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } } }
    delete:
      security: [ { bearerAuth: [] } ]
      parameters:
        - { name: tenant, in: path, required: true, schema: { type: string } }
        - { name: name, in: path, required: true, schema: { type: string } }
      responses: { '200': { description: ok } }
    parameters:
      - { name: tenant, in: path, required: true, schema: { type: string } }
      - { name: name, in: path, required: true, schema: { type: string } }
  /api/v1/apps:
    post:
      security: [ { bearerAuth: [] } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/App' } } } }
      responses: { '200': { description: ok } }
  /api/v1/apps/{tenant}/{project}/{name}:
    get:
      security: [ { bearerAuth: [] } ]
      parameters:
        - { name: tenant, in: path, required: true, schema: { type: string } }
        - { name: project, in: path, required: true, schema: { type: string } }
        - { name: name, in: path, required: true, schema: { type: string } }
      responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/App' } } } } }
    delete:
      security: [ { bearerAuth: [] } ]
      parameters:
        - { name: tenant, in: path, required: true, schema: { type: string } }
        - { name: project, in: path, required: true, schema: { type: string } }
        - { name: name, in: path, required: true, schema: { type: string } }
      responses: { '200': { description: ok } }
  /api/v1/kubeconfig-grants:
    post:
      security: [ { bearerAuth: [] } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/KubeconfigGrant' } } } }
      responses:
        '200': { description: kubeconfig issued, content: { application/json: { schema: { $ref: '#/components/schemas/KubeconfigGrant' } } } }
  /api/v1/clusters:
    post:
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                kubeconfig: { type: string, format: byte }
                labels: { type: object, additionalProperties: { type: string } }
              required: [name, kubeconfig]
      responses:
        '200': { description: cluster registered, content: { application/json: { schema: { $ref: '#/components/schemas/Tenant' } } } }
  /api/v1/clusters/{id}:
    get:
      security: [ { bearerAuth: [] } ]
      parameters: [ { name: id, in: path, required: true, schema: { type: integer } } ]
      responses:
        '200': { description: cluster status, content: { application/json: { schema: { type: object } } } }
  /api/v1/clusters/{id}/events:
    get:
      security: [ { bearerAuth: [] } ]
      parameters: [ { name: id, in: path, required: true, schema: { type: integer } } ]
      responses:
        '200': { description: events, content: { application/json: { schema: { type: array, items: { type: object } } } } }